requirements.txt
nicegui
openai
python-dotenv
pytest